add ploopy mini keymap

+++ sync.sh
@@ -12,7 +12,9 @@ sync_files()
     src="$1"
     dest="$2"
     printf 'synchronizing %s\n' "$1"
-    rm -r "$dest"
+    if [ -d "$dest" ]; then
+        rm -r "$dest"
+    fi
     cp -arT "$src" "$dest"
 }
 
@@ -21,6 +23,7 @@ if [ -d "${qmk_home:?}" ]; then
     sync_files "userspace" "${qmk_home:?}/users/f8"
     sync_files "4x12" "${qmk_home:?}/keyboards/handwired/4x12"
     sync_files "lily58" "${qmk_home:?}/keyboards/lily58/keymaps/f8"
+    sync_files "ploopy_mini" "${qmk_home:?}/keyboards/ploopyco/trackball_mini/keymaps/f8"
 fi
 
 printf 'done.\n'
